Add the boost package.
This also fixes some of the other config scripts so that they don't unpack
the tar files everything they run.  This means that the packages don't get
rebuilt everytime (a good thing).

+++ LCG_Builders/dawn/scripts/dawn_config.sh
@@ -1,11 +1,19 @@
 #!/bin/sh
 
+package_directory=dawn_${LCG_package_config_version}
+
 cd ${LCG_builddir}
+if [ -d ${package_directory} ]; then
+    echo XXXX ${LCG_builddir}/${package_directory} already exists. 
+    echo XXXX Not unpacking and reconfiguring.  
+    echo XXXX Remove directory to force reconfiguration.  
+    exit 0
+fi
+
 tar xvfz ${LCG_tarfilename}
+cd ${package_directory}
 
-cd dawn_${LCG_package_config_version}
 rm dev_control.h
-
 cat > dev_control.h << EOF
 
 #if !defined DEV_CONTROL_H
